Huge wave of changes involving the archetecture of loading historic data files.
FilesystemHelpers became a static class
CSV file finding and parsing moved into FilesystemHelpers.
BehavioralBoxHistoricalData became less autonomous and more of a wrapper and event parser for a LabjackDataFile vector.
BehavioralBoxControllersManager became more controlling, and it holds a map of labjackSerialNumber ints to vectors of LabjackDataFiles. These are what are used to build BehavioralBoxHistoricalData objects (which is still stores a vector of).